TIP-874: move factory related classes
