+++ src/Akeneo/Pim/Enrichment/Bundle/Resources/config/factories.yml
@@ -1,18 +1,18 @@
 parameters:
-    pim_catalog.factory.metric.class:                Pim\Component\Catalog\Factory\MetricFactory
-    pim_catalog.factory.price.class:                 Pim\Component\Catalog\Factory\PriceFactory
-    pim_catalog.factory.group.class:                 Pim\Component\Catalog\Factory\GroupFactory
-    pim_catalog.factory.product_unique_data.class:   Pim\Component\Catalog\Factory\ProductUniqueDataFactory
+    pim_catalog.factory.metric.class:                Akeneo\Pim\Enrichment\Component\Product\Factory\MetricFactory
+    pim_catalog.factory.price.class:                 Akeneo\Pim\Enrichment\Component\Product\Factory\PriceFactory
+    pim_catalog.factory.group.class:                 Akeneo\Pim\Enrichment\Component\Product\Factory\GroupFactory
+    pim_catalog.factory.product_unique_data.class:   Akeneo\Pim\Enrichment\Component\Product\Factory\ProductUniqueDataFactory
 
-    pim_catalog.factory.value_collection.class:       Pim\Component\Catalog\Factory\ValueCollectionFactory
-    pim_catalog.factory.value.class:                  Pim\Component\Catalog\Factory\ValueFactory
-    pim_catalog.factory.value.scalar.class:           Pim\Component\Catalog\Factory\Value\ScalarValueFactory
-    pim_catalog.factory.value.metric.class:           Pim\Component\Catalog\Factory\Value\MetricValueFactory
-    pim_catalog.factory.value.price_collection.class: Pim\Component\Catalog\Factory\Value\PriceCollectionValueFactory
-    pim_catalog.factory.value.option.class:           Pim\Component\Catalog\Factory\Value\OptionValueFactory
-    pim_catalog.factory.value.options.class:          Pim\Component\Catalog\Factory\Value\OptionsValueFactory
-    pim_catalog.factory.value.media.class:            Pim\Component\Catalog\Factory\Value\MediaValueFactory
-    pim_catalog.factory.value.date.class:             Pim\Component\Catalog\Factory\Value\DateValueFactory
+    pim_catalog.factory.value_collection.class:       Akeneo\Pim\Enrichment\Component\Product\Factory\ValueCollectionFactory
+    pim_catalog.factory.value.class:                  Akeneo\Pim\Enrichment\Component\Product\Factory\ValueFactory
+    pim_catalog.factory.value.scalar.class:           Akeneo\Pim\Enrichment\Component\Product\Factory\Value\ScalarValueFactory
+    pim_catalog.factory.value.metric.class:           Akeneo\Pim\Enrichment\Component\Product\Factory\Value\MetricValueFactory
+    pim_catalog.factory.value.price_collection.class: Akeneo\Pim\Enrichment\Component\Product\Factory\Value\PriceCollectionValueFactory
+    pim_catalog.factory.value.option.class:           Akeneo\Pim\Enrichment\Component\Product\Factory\Value\OptionValueFactory
+    pim_catalog.factory.value.options.class:          Akeneo\Pim\Enrichment\Component\Product\Factory\Value\OptionsValueFactory
+    pim_catalog.factory.value.media.class:            Akeneo\Pim\Enrichment\Component\Product\Factory\Value\MediaValueFactory
+    pim_catalog.factory.value.date.class:             Akeneo\Pim\Enrichment\Component\Product\Factory\Value\DateValueFactory
 
 services:
     pim_catalog.factory.family_variant:
