+++ src/Akeneo/Pim/Enrichment/Bundle/spec/Doctrine/Common/Saver/ProductUniqueDataSynchronizerSpec.php
@@ -6,7 +6,7 @@ use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
 use PhpSpec\ObjectBehavior;
 use Akeneo\Pim\Enrichment\Bundle\Doctrine\Common\Saver\ProductUniqueDataSynchronizer;
-use Pim\Component\Catalog\Factory\ProductUniqueDataFactory;
+use Akeneo\Pim\Enrichment\Component\Product\Factory\ProductUniqueDataFactory;
 use Akeneo\Pim\Structure\Component\Model\AttributeInterface;
 use Akeneo\Pim\Enrichment\Component\Product\Model\ProductInterface;
 use Akeneo\Pim\Enrichment\Component\Product\Model\ProductUniqueDataInterface;
