+++ src/Akeneo/Pim/Enrichment/Bundle/spec/EventSubscriber/LoadEntityWithValuesSubscriberSpec.php
@@ -5,7 +5,7 @@ namespace spec\Akeneo\Pim\Enrichment\Bundle\EventSubscriber;
 use Doctrine\Common\Persistence\Event\LifecycleEventArgs;
 use PhpSpec\ObjectBehavior;
 use Akeneo\Pim\Enrichment\Bundle\EventSubscriber\LoadEntityWithValuesSubscriber;
-use Pim\Component\Catalog\Factory\ValueCollectionFactoryInterface;
+use Akeneo\Pim\Enrichment\Component\Product\Factory\ValueCollectionFactoryInterface;
 use Akeneo\Pim\Enrichment\Component\Product\Model\ProductInterface;
 use Pim\Component\Catalog\Model\ValueCollectionInterface;
 use Prophecy\Argument;
