+++ src/Akeneo/Pim/Enrichment/Component/Product/Builder/EntityWithValuesBuilder.php
@@ -5,7 +5,7 @@ namespace Akeneo\Pim\Enrichment\Component\Product\Builder;
 use Akeneo\Pim\Enrichment\Component\Product\Model\ProductInterface;
 use Akeneo\Pim\Structure\Component\Model\AttributeInterface;
 use Pim\Component\Catalog\AttributeTypes;
-use Pim\Component\Catalog\Factory\ValueFactory;
+use Akeneo\Pim\Enrichment\Component\Product\Factory\ValueFactory;
 use Akeneo\Pim\Enrichment\Component\Product\Manager\AttributeValuesResolverInterface;
 use Pim\Component\Catalog\Model\EntityWithValuesInterface;
