+++ src/Akeneo/Pim/Enrichment/Component/Product/Factory/ValueFactory.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace Pim\Component\Catalog\Factory;
+namespace Akeneo\Pim\Enrichment\Component\Product\Factory;
 
 use Akeneo\Pim\Enrichment\Component\Product\Model\ValueInterface;
 use Akeneo\Pim\Structure\Component\Model\AttributeInterface;
 use Pim\Component\Catalog\Exception\InvalidAttributeException;
-use Pim\Component\Catalog\Factory\Value\ValueFactoryInterface;
+use Akeneo\Pim\Enrichment\Component\Product\Factory\Value\ValueFactoryInterface;
 use Pim\Component\Catalog\Validator\AttributeValidatorHelper;
 
 /**
