+++ src/Akeneo/Pim/Enrichment/Component/spec/Product/Factory/Value/ScalarValueFactorySpec.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace spec\Pim\Component\Catalog\Factory\Value;
+namespace spec\Akeneo\Pim\Enrichment\Component\Product\Factory\Value;
 
 use Akeneo\Tool\Component\StorageUtils\Exception\InvalidPropertyTypeException;
 use PhpSpec\ObjectBehavior;
-use Pim\Component\Catalog\Factory\Value\ScalarValueFactory;
+use Akeneo\Pim\Enrichment\Component\Product\Factory\Value\ScalarValueFactory;
 use Akeneo\Pim\Structure\Component\Model\AttributeInterface;
 use Akeneo\Pim\Enrichment\Component\Product\Value\ScalarValue;
 use Prophecy\Argument;
